ChannelOpener WIP
