+++ btc-lsp/src/BtcLsp/Storage/Model/LnChan.hs
@@ -19,11 +19,35 @@ createIgnore ::
   m (Entity LnChan)
 createIgnore swapId txid vout ss = runSql $ do
   ct <- getCurrentTime
+  Psql.update $ \swap -> do
+    Psql.set
+      swap
+      --
+      -- TODO : better status mapping
+      --
+      [ SwapIntoLnStatus
+          Psql.=. Psql.val SwapWaitingChan,
+        SwapIntoLnUpdatedAt
+          Psql.=. Psql.val ct
+      ]
+    Psql.where_ $
+      ( swap Psql.^. SwapIntoLnId
+          Psql.==. Psql.val swapId
+      )
   Psql.upsertBy
     (UniqueLnChan txid vout)
-    (this ct)
+    LnChan
+      { lnChanSwapIntoLnId = swapId,
+        lnChanFundingTxId = txid,
+        lnChanFundingVout = vout,
+        lnChanClosingTxId = Nothing,
+        lnChanNumUpdates = 0,
+        lnChanStatus = ss,
+        lnChanInsertedAt = ct,
+        lnChanUpdatedAt = ct
+      }
     --
-    -- TODO : this update is redundant, but upsertBy is
+    -- TODO : txid + vout update is redundant, but upsertBy is
     -- not working with mempty update argument -
     -- probably it's a bug in Esqueleto implementation,
     -- check it in latest version, and if not fixed -
@@ -33,20 +57,9 @@ createIgnore swapId txid vout ss = runSql $ do
     -- https://github.com/bitemyapp/esqueleto/issues/294
     --
     [ LnChanFundingTxId Psql.=. Psql.val txid,
-      LnChanFundingVout Psql.=. Psql.val vout
+      LnChanFundingVout Psql.=. Psql.val vout,
+      LnChanSwapIntoLnId Psql.=. Psql.val swapId
     ]
-  where
-    this ct0 =
-      LnChan
-        { lnChanSwapIntoLnId = swapId,
-          lnChanFundingTxId = txid,
-          lnChanFundingVout = vout,
-          lnChanClosingTxId = Nothing,
-          lnChanNumUpdates = 0,
-          lnChanStatus = ss,
-          lnChanInsertedAt = ct0,
-          lnChanUpdatedAt = ct0
-        }
 
 getByChannelPoint ::
   (Env m) =>
