+++ btc-lsp/src/BtcLsp/Thread/Main.hs
@@ -8,8 +8,9 @@ where
 
 import BtcLsp.Data.AppM (runApp)
 import BtcLsp.Import
-import qualified BtcLsp.Storage.Migration as StorageMigration
-import qualified BtcLsp.Thread.Server as ThreadServer
+import qualified BtcLsp.Storage.Migration as Storage
+import qualified BtcLsp.Thread.ChannelOpener as ChannelOpener
+import qualified BtcLsp.Thread.Server as Server
 import qualified LndClient.RPC.Katip as Lnd
 
 main :: IO ()
@@ -23,11 +24,12 @@ apply = do
   unlocked <- withLnd Lnd.lazyUnlockWallet id
   if isRight unlocked
     then do
-      StorageMigration.migrateAll
+      Storage.migrateAll
       xs <-
         mapM
           spawnLink
-          [ ThreadServer.apply
+          [ Server.apply,
+            ChannelOpener.apply
           ]
       liftIO
         . void
