--- btc-lsp/src/BtcLsp/Thread/SwapIntoLn.hs
@@ -1,69 +1,0 @@
-{-# LANGUAGE TemplateHaskell #-}
-
-module BtcLsp.Thread.SwapIntoLn
-  ( apply,
-  )
-where
-
-import BtcLsp.Import
-import qualified BtcLsp.Storage.Model.SwapIntoLn as SwapIntoLn
-import qualified Data.Set as Set
-import qualified LndClient.Data.OpenChannel as Chan
-import qualified LndClient.Data.Peer as Peer
-import qualified LndClient.RPC.Katip as Lnd
-
-apply :: (Env m) => m ()
-apply = do
-  peerList <- fromRight [] <$> withLnd Lnd.listPeers id
-  let peerSet = Set.fromList $ Peer.pubKey <$> peerList
-  swaps <- SwapIntoLn.getFundedSwaps
-  tasks <-
-    mapM
-      ( spawnLink
-          . openChan
-      )
-      $ filter
-        ( \x ->
-            Set.member
-              (userNodePubKey . entityVal $ snd x)
-              peerSet
-        )
-        swaps
-  mapM_ (liftIO . wait) tasks
-  sleep $ MicroSecondsDelay 500000
-  apply
-
-openChan :: (Env m) => (Entity SwapIntoLn, Entity User) -> m ()
-openChan (swapEnt, userEnt) = do
-  res <- runExceptT $ do
-    --
-    -- TODO : !!!
-    --
-    withLndT
-      Lnd.openChannelSync
-      ( $
-          Chan.OpenChannelRequest
-            { Chan.nodePubkey =
-                userNodePubKey $
-                  entityVal userEnt,
-              Chan.localFundingAmount =
-                --
-                -- TODO : correct cap
-                --
-                from . swapIntoLnChanCapLsp $
-                  entityVal swapEnt,
-              Chan.pushMSat = Nothing,
-              Chan.targetConf = Nothing,
-              Chan.mSatPerByte = Nothing,
-              Chan.private = Nothing,
-              Chan.minHtlcMsat = Nothing,
-              Chan.remoteCsvDelay = Nothing,
-              Chan.minConfs = Nothing,
-              Chan.spendUnconfirmed = Nothing,
-              Chan.closeAddress = Nothing
-            }
-      )
-  whenLeft res $
-    $(logTM) ErrorS . logStr
-      . ("OpenChan procedure failed: " <>)
-      . inspect
